web_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
tensorflow==2.15.0
chinesecalendar==1.8.0
